refactor: init
- 재설계 초기상태
